backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
finnhub-python==2.4.19
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pydantic==2.5.0
neuralprophet==0.7.0
prophet==1.1.5
torch==2.1.1
python-multipart==0.0.6
python-dotenv==1.0.0